requirements.txt
alembic==1.8.1
aniso8601==9.0.1
attrs==22.1.0
autopep8==2.0.0
cffi==1.15.1
click==8.1.3
colorama==0.4.6
cryptography==38.0.3
exceptiongroup==1.0.3
Flask==2.2.2
Flask-Migrate==3.1.0
Flask-RESTful==0.3.9
Flask-SQLAlchemy==3.0.2
greenlet==2.0.1
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.3
MarkupSafe==2.1.1
packaging==21.3
pluggy==1.0.0
pycodestyle==2.9.1
pycparser==2.21
PyJWT==2.6.0
pyparsing==3.0.9
pytest==7.2.0
pytz==2022.6
six==1.16.0
SQLAlchemy==1.4.43
tomli==2.0.1
Werkzeug==2.2.2
